[JBMB-48] 진단 기록 삭제 API 구현 (이한범)

+++ jbmb_coreserver/src/main/java/com/jbmb/jbmb_coreserver/diagnosis/repository/UpdateLogRepository.java
@@ -10,12 +10,11 @@ import javax.persistence.criteria.CriteriaBuilder;
 import java.util.Optional;
 
 @Repository
-public interface UpdateLogRepository extends JpaRepository<DiagnosisLog, Long> {
+public interface UpdateLogRepository extends JpaRepository<DiagnosisLog, Integer> {
 
     @Query("select l.id from DiagnosisLog l where l.userNum=?1 and l.active=0")
     Integer findLogByUserNum(Integer userNum);
-    // active 
+    // active
     // 0:설문조사 진행 중
-    // 1:설문조사 완료 및 이미지 미제출
-    // 2:이미지 제출
+    // 1:설문조사 완료
 }
